Learning linked lists, modified the atoi because it has to accept INT_MAX/INT_MIN

+++ push_swap/utils.c
@@ -1,9 +1,9 @@
 #include "push_swap.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int					neg;
-	unsigned long int	result;
+	unsigned long		result;
 	size_t				index;
 
 	neg = 1;
